refactor: Semantically ingestion happens into staging
