+++ src/handlers/http/modal/ingest/ingestor_logstream.rs
@@ -34,7 +34,7 @@ use crate::{
     },
     metadata,
     option::CONFIG,
-    staging::STREAM_WRITERS,
+    staging::STAGING,
     stats,
 };
 
@@ -75,7 +75,7 @@ pub async fn delete(stream_name: Path<String>) -> Result<impl Responder, StreamE
     }
 
     metadata::STREAM_INFO.delete_stream(&stream_name);
-    STREAM_WRITERS.delete_stream(&stream_name);
+    STAGING.delete_stream(&stream_name);
     stats::delete_stats(&stream_name, "json")
         .unwrap_or_else(|e| warn!("failed to delete stats for stream {}: {:?}", stream_name, e));
