+++ src/handlers/http/modal/query/querier_logstream.rs
@@ -47,7 +47,7 @@ use crate::{
     hottier::HotTierManager,
     metadata::{self, STREAM_INFO},
     option::CONFIG,
-    staging::{Staging, STREAM_WRITERS},
+    staging::{Stream, STAGING},
     stats::{self, Stats},
     storage::StreamType,
 };
@@ -69,7 +69,7 @@ pub async fn delete(stream_name: Path<String>) -> Result<impl Responder, StreamE
     let objectstore = CONFIG.storage().get_object_store();
 
     objectstore.delete_stream(&stream_name).await?;
-    let stream_dir = Staging::new(&CONFIG.options, &stream_name);
+    let stream_dir = Stream::new(&CONFIG.options, &stream_name);
     if fs::remove_dir_all(&stream_dir.data_path).is_err() {
         warn!(
             "failed to delete local data for stream {}. Clean {} manually",
@@ -102,7 +102,7 @@ pub async fn delete(stream_name: Path<String>) -> Result<impl Responder, StreamE
     }
 
     metadata::STREAM_INFO.delete_stream(&stream_name);
-    STREAM_WRITERS.delete_stream(&stream_name);
+    STAGING.delete_stream(&stream_name);
     stats::delete_stats(&stream_name, "json")
         .unwrap_or_else(|e| warn!("failed to delete stats for stream {}: {:?}", stream_name, e));
