+++ src/metadata.rs
@@ -37,7 +37,7 @@ use crate::metrics::{
     LIFETIME_EVENTS_INGESTED_SIZE,
 };
 use crate::option::CONFIG;
-use crate::staging::Staging;
+use crate::staging::Stream;
 use crate::storage::retention::Retention;
 use crate::storage::{ObjectStorage, ObjectStoreFormat, StreamType};
 use crate::utils::arrow::MergedRecordReader;
@@ -370,7 +370,7 @@ impl StreamInfo {
 }
 
 fn update_schema_from_staging(stream_name: &str, current_schema: Schema) -> Schema {
-    let staging_files = Staging::new(&CONFIG.options, stream_name).arrow_files();
+    let staging_files = Stream::new(&CONFIG.options, stream_name).arrow_files();
     let record_reader = MergedRecordReader::try_new(&staging_files).unwrap();
     if record_reader.readers.is_empty() {
         return current_schema;
