+++ src/storage/object_storage.rs
@@ -31,7 +31,7 @@ use crate::handlers::http::users::{CORRELATION_DIR, DASHBOARDS_DIR, FILTER_DIR, 
 use crate::metadata::SchemaVersion;
 use crate::metrics::{EVENTS_STORAGE_SIZE_DATE, LIFETIME_EVENTS_STORAGE_SIZE};
 use crate::option::Mode;
-use crate::staging::{convert_disk_files_to_parquet, Staging};
+use crate::staging::{convert_disk_files_to_parquet, Stream};
 use crate::{
     alerts::Alerts,
     catalog::{self, manifest::Manifest, snapshot::Snapshot},
@@ -551,7 +551,7 @@ pub trait ObjectStorage: Debug + Send + Sync + 'static {
             let custom_partition = STREAM_INFO
                 .get_custom_partition(stream)
                 .map_err(|err| ObjectStorageError::UnhandledError(Box::new(err)))?;
-            let dir = Staging::new(&CONFIG.options, stream);
+            let dir = Stream::new(&CONFIG.options, stream);
             let schema = convert_disk_files_to_parquet(
                 stream,
                 &dir,
