+++ src/sync.rs
@@ -24,7 +24,7 @@ use tokio::time::{interval, sleep, Duration};
 use tracing::{error, info, warn};
 
 use crate::option::CONFIG;
-use crate::staging::STREAM_WRITERS;
+use crate::staging::STAGING;
 use crate::{storage, STORAGE_UPLOAD_INTERVAL};
 
 pub async fn object_store_sync() -> (
@@ -98,7 +98,7 @@ pub async fn run_local_sync() -> (
             scheduler
                 .every((storage::LOCAL_SYNC_INTERVAL as u32).seconds())
                 .run(|| async {
-                    STREAM_WRITERS.unset_all();
+                    STAGING.unset_all();
                 });
 
             loop {
